+++ src/command-line/parsing/parser.hpp
@@ -5,20 +5,8 @@
 
 namespace found {
 
-/**
- * This is where the program starts.
- * 
- * @param argc The number of arguments passed into the command line
- * @param argv The arguments passed into the command line
- * 
- * @return An integer indicating success (0) iff the program executes
- * successfully
- * 
- * @note The method itself uses command line arguments to generate
- * an Options object that represents all the algorithms we want to run
- * and their parameters
-*/
-int main(int argc, char **argv);
+/// Defines the help message
+#define HELP_MSG "Use ./found --help or ./found -h for help"
 
 /**
  * Parses the calibration options from the command line to
